LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/css/css-font-loading/nonexistent-file-url-expected.txt

CONSOLE MESSAGE: Not allowed to load local resource: thisisnotarealfile.otf

PASS nonexistent-file-url
